requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Faker==8.12.0
freezegun==1.1.0
model-bakery==1.3.2
parameterized==0.8.1
Pillow==8.3.2
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
